Fix the VCR
