+++ spec/dummy/spec/support/test_helpers.rb
@@ -1,0 +1,4 @@
+def restore_default_config
+  Planter.configuration = nil
+  Planter.configure {}
+end
