--- test/planter_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class Planter::Test < ActiveSupport::TestCase
-  test "truth" do
-    assert_kind_of Module, Planter
-  end
-end
